requirements/prod.txt
# requirements/prod.txt
-r common.txt
django-storages==1.1.8
boto==2.17.0
